refactor: extract render context classes
